+++ apps/nuxt/components/blocks/product/ProductDetails.vue
