+++ apps/nuxt/components/blocks/product/ProductVerticalGallery.vue
